fix: adicionar validação de UUID em operações de CRUD

+++ src/users/users.service.ts
@@ -4,6 +4,7 @@ import { UpdateUserDto } from './dto/update-user.dto';
 import { Repository } from 'typeorm';
 import { User } from './entities/user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
+import { isUUID } from 'class-validator';
 
 @Injectable()
 export class UsersService {
@@ -27,20 +28,36 @@ export class UsersService {
   }
 
   async findOne(id: string): Promise<User> {
+    if (!isUUID(id)) {
+      throw new NotFoundException(`Usuário com ID ${id} não encontrado`);
+    }
     const user = await this.userRepository.findOne({ where: { id } });
     if (!user) {
-      throw new NotFoundException(`User with id "${id}" not found`);
+      throw new NotFoundException(`Usuário com ID ${id} não encontrado`);
     }
     return user;
   }
 
   async update(id: string, updateUserDto: UpdateUserDto): Promise<User> {
-    const user = await this.findOne(id);
+    if (!isUUID(id)) {
+      throw new NotFoundException(`Usuário com ID ${id} não encontrado`);
+    }
+
+    const user = await this.userRepository.findOne({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException(`Usuário com ID ${id} não encontrado`);
+    }
+
     Object.assign(user, updateUserDto);
+
     return await this.userRepository.save(user);
   }
 
   async remove(id: string): Promise<void> {
+    if (!isUUID(id)) {
+      throw new NotFoundException(`Usuário com ID ${id} não encontrado`);
+    }
     const result = await this.userRepository.delete(id);
     if (result.affected === 0) {
       throw new NotFoundException(`User with id "${id}" not found`);
